refactor(web): split sessions data helpers [OPE-345]
